Moved sharedContentModal
