+++ resources/assets/homepageCard.js
